Ryan Wilkie and Venkat Worked on shooter: pivot, flywheels, and feeder
added PID's to them, added their values to smartdashboard, created constants class for pivot positions
(for amp, trap, and speaker) they are NOT correct and will need tuning later. Also added boolean forward parameter for pivot and feeder. We also made it so if the limit switch inside holder is tripped, feeder stops.